+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.0.0
+
+- Require SDK 3.3.0+.
+- Add class modifiers.
+- Update lints.
+
+- Change some `void` return types to `Null` in `JsonReader`
+  This allows uses like `reader.tryString() ?? reader.expectNull()`
+  to match a string or `null`.
+- Add `processObjectEntry` method to `JsonProcessor`, called
+  for each object entry.
+- Use table to improve parsing speed for `skipAnyValue`.
+- Faster whitespace skipping in readers.
+
 ## 1.2.0
 
 - Allow SDK 3.0.0+.
@@ -18,6 +32,8 @@
 
 - Populate the pubspec's `repository` field.
 - Use `package:lints` for linting.
+- Adds `processObjectEntry` to `JsonProcessor`.
+
 ## 1.1.0
 
 - Null safe.